tester/requirements.txt
pandas
scikit-learn
tensorflow-cpu
matplotlib
seaborn